+++ javai-persistence/src/test/java/dev/xtrafe/javai/persistence/scanned/ScannedOnlyEntityRepository.java
@@ -1,0 +1,6 @@
+package dev.xtrafe.javai.persistence.scanned;
+
+import dev.xtrafe.javai.persistence.JavAIRepository;
+
+public interface ScannedOnlyEntityRepository extends JavAIRepository<ScannedOnlyEntity> {
+}